Added script to recreate the database schema.
This script uses DBIx::Class::Schema::Loader to create Schema Result classes
for the papsdb database.  It uses a number of settings (components, use_moose,
etc.) to make the output match as much as possible the previous schema that
was used and was created by the Catalyst helper scripts.  One difference is to
use a newer naming convention; namely, this script uses v7, while the previous
schema used v4.  The primary difference to the inflect the table names so the
created classes have singular names instead of plural names.