Moved inserting db records into a new sql file.
The paps.sql file previously contained both the sql to create the database
tables and the sql to populate them with default or sample records.  Moved
populating these tables with records to a new file.

Currently, however, both scripts need to be run in tandem, as the
paps-data.sql script uses the primary key values to reference records in other
tables; if the data is repopulated but the tables are not recreated, the
autoincrementing fields will have different values, causing a foreign key
violation.