Updated schema to DBIx-Class-Schema-Loader-0.07015.
Recreated the schema for the papsdb using the latest version (0.07015) of
DBIx::Class::Schema::Loader.  The most notable change in this version is being
able to singularize "People" to "Person", so the model class for the people
table is now Person.pm.  Also, the many-to-many relationships are now
automatically created, but, since I manually created these already, the new
relationships should be overridden by the manual ones.