Added smallint column 'year' to works table.
Added a new column to the works table, a 'year' column with type smallint.
This column is meant to contain the year a work was released.  PostgreSQL does
not have unsigned types, so this is simply a smallint, which holds values up
to +32767, more than enough for any plausible year.