Redefined the password_hash column with encoding information.
The password_hash column is an encoded column, using a SHA-1 hash algorithm.
The Result class that was automatically generated does not include this
information, so we redefine the column here to also have those properties.
Note that, while the settings include a salt lenght, I am not sure if the
column is being hashed with a salt; this needs further investigation (the
encode class might need to be modified).