Set the created date to be retrieved on insert.
The created_date field defaults to the current time, so if this is not
provided when a record (a user) is created, it should be retrieved.  This
setting, of course, was not set by default when this Result class was
created, so the column is redefined here to include that setting.