Recreated Result classes to include new rank column.
The new rank column in the source_work_categories table is now part of the
SourceWorkCategory result class.  Apparently, as well, now that that table has
more columns than what is in its primary key, the DBIx::Class::Schema::Loader
does not want to automatically create many-to-many relationships between the
Work and SourceCategory classes.  This is fine, as I have already manually
defined these relationships.  I might need to tweak them, however, to bring
along the rank column when appropriate.