Added new author metadata columns to work_authors table.
Added two new columns to the work_authors table.  The author_name_text column
optionally stores the literal text used in the work to name an author.  This
can be used to keep track of how an author is named and to store a name for an
author whose definitive "person" information has not yet been found.  The
second new column, author_affiliation_text, optionally stores the literal text
used to describe the author's affiliation.  Since there is no definitive way
to store affiliation information, this will be the only way, for now, to keep
track of an author's affiliation.  Later, like the author_name_text column, it
will be used to keep track of how an affiliation is named and to keep track of
affiliations that have not be definitively identified yet.