Updated result classes for collection and collection_works tables.
Added new result classes Collection and CollectionWork, for the new tables
collections and collection_works, respectively.  There is now also a one-to-
many relationship between the users and collections table, resulting in a new
has_many relationship in the User result class; and there is a new many-to-
many relationship between works and collections, where collection_works is the
linking table.