Added many-to-many relationship between collections and works.
The DBIx::Class::Schema::Loader module (as of version 0.07015) only
automatically creates many-to-many relationships if the linking tables for
those relationships are comprised only of the foreign keys relating the two
tables in the relationships.  The collection_works table has additional
metadata fields (added_timestamp and notes), so the many-to-many relationship
must be defined manually.  Also, we want the metadata values included in the
result sets for the linked objects, so the relationship would need to be
redefined, anyway.