Added missing columns to the view's select clause.
The initial version of this view only returned a few columns, specifically the
columns used while testing the query.  This query should return all (or, at
least, most) of the columns from the works table, as well as some columsn from
the work_references table.  The select statements inside the CTE and the outer
select statement have been updated to return these values.