Re-added the relationships from the Work result set.
This view, WorkToRead, has at least all of the columns from the works table,
so it is possible it will make use of all of its relationships.  Added to this
view each of the relationships in the Work result set.  However, corresponding
relationships have not been added to the related result sets, so they would
need to be added later if the reciprocal relationships are needed.