Added view to return all works with user data and references.
Added view ViewWorksWithUserData.  This view returns a list of all works, along
with a specific user's data for each work (or null, if no user is specified).
Each work also has all works that reference it, so a work referenced multiple
times will be returned multiple times.

This view is used instead of simply using the Work result set's
user_work_datas relationship, as that will return no rows for a work if the
indicated user does not have any metadata for that work (no record in
user_work_data), but at least one other user does have metadata for that work.