Added result classes Algorithm and ReferencedWorkGuess.
The Algorithm result class relates to the algorithms table, and is simply a
lookup table.  The ReferencedWorkGuess result class relates to the
referenced_work_guesses table.  This holds all of the guesses each uses has
made with an algorithm for what a work reference is referencing.  Relationship
were added (in the previous commit) in other classes, namely User and
WorkReference classes.  The Algorithm class also has a relationship to this
class.