Added user to referenced work guess unique key.
The unique constraint in the referenced_work_guesses table now includes the
user_id column, as well.  This means two different users can use the same
algorithm -- and version of that algorithm -- to make two different guesses.
For most automated situations, this will not be an issue, as if there are two
different users making a guess on a reference, they will use different
algorithms (or versions thereof).  If two are using the same algorithm, there
is probably no reason to both guess on the same references.  However, this
can most likely easily be covered in the application layer.

The important part here is if we have humans make manual guesses on
references.  Right now, users can change the actual referenced work.  But we
might want to make this a two-step process, where people put in guesses, and
only with a certain confidence, or approval, does the guess change the actual
reference.  The guesses people make would have the same "algorithm," and thus
would violate the previous version of the unique constraint.