Use work_reference_id, not guessed_referenced_work_id, in unique constraint.
The referenced_work_guesses table's unique constraint was incorrect.  It was
acting as if a persona could only guess a work was the one being referenced
once.  Instead, it should only be providing one guess per work reference.
There is not problem if a persona thinks a work is being referenced by
multiple references.